Refactored, added to README, ready loop can die
During a worker wrappers loop the routine wouldn't end until the worker
answers yes to Ready, this is unintuitive so I've added a Uint32 that
can be atomically checked after each Ready call. If the worker answers
no but the pool is closed then we escape the loop.

This loop still needs looking at since there's still no guarantee that
the go routine is stopped when pool.Close finishes, there should be
something like a join() to ensure this, maybe.