Added TunnyInterupt call for custom workers
TunnyInterupt is an optional func that workers may implement which is
called when a job has timed out. It indicates that there is no longer a
client on the other end of this job, and it can be discarded (return
nil).

The call is made on an outside goroutine and you need to be careful how
this is implemented.