TunnyWorker interface changed
This is a breaking change. I've renamed Job and Ready in the TunnyWorker
interface to TunnyJob and TunnyReady respectively in order to avoid
collisions with other interfaces.

I've also removed the tests for counting goroutines since they were
flaky to begin with, but were always failing with go test -race.