document presentasione..
Matt Lee, a lead developer of The List, which is licensed under the <a href="https://www.gnu.org/licenses/agpl.html">GNU Affero General Public License version 3</a> (AGPLv3), or at your option, any later version. Matt is the technical le
]]>
<![CDATA[
ad at Creative Commons. Matt has been working in free software for over a decade and is a notable contributor to the GNU project and a former Campaigns Manager at the Free Software Foundation, as well as co-founder of Libre.fm and GNU social. Currently Matt is <a href="http://indiegogo.com/at/orangu">raising funds</a> to produce a film this summer, <em>Orang-U: An Ape Goes to College</em>, which he plans to edit using entirely free software and release under a <a href="https://creativecommons.org/licenses/by-sa/4.0/">CC BY-SA</a> license.</p> <h2>Can you tell us a bit about The List?</h2> <p>No one can be everywhere at once. But everyone can.</p> <p>NGOs, journalists, government agencies, and cultural institutions all need photographs to tell their story and educate others. But there’s no way for those organizations to be in the right place at the right time, every time. That’s where we come in.</p> <p>Through The List, organizations will provide lists of locations, people, and events that they need photographs of. And when users are in the right place at the right time, they can claim an item from the list and publish a photograph of it.</p> <h2>What inspired the creation of The List?</h2> <p>The List powered by Creative Commons is an experiment to see if we can make it easier for people to contribute to the public commons. There are millions of places for images that exist in the public commons in our daily lives, from newspaper articles to photos and illustrations on Wikipedia. The List hopes to bring the people who have a need for such images and the people who may take them, together.</p> <h2>How are people using it?</h2> <p>Right now we're still prototyping things and we're working on getting some real world items into The List, but the way it works is pretty simple: you fire up the app, choose some categories of things you're interested in such as pets, beverages, etc. Then we show you an item that matches the category, and the requesting organization (Wikipedia, FSF, etc.) can choose to add it to your list or not. After you go through this process a couple of times, you wind up with a personal to-do list of images. Taking an image is easy, too: just tap it, tap the camera and take your photo. Or you can upload a photo you've already taken. The photo is then sent to some servers at Creative Commons, where we add metadata and produce a variety of thumbnails of the image, before sending it over to the Internet Archive for permanent storage. And moments later, the image is in the gallery in the app on your phone.</p> <h2>What features do you think really sets The List apart from similar software?</h2> <p>I don't think there's anything else like this out there. For the first time, there's an application that makes it quick and easy to contribute to the public commons, and we do that by hiding a lot of the detail away from the user. For example, instead of presenting a choice of the six Creative Commons licenses, we choose one and all images are licensed in the same manner. It's also a license that's compatible with Wikipedia and other similar projects—Creative Commons Attribution 4.0.</p> <h2>Why did you choose the AGPLv3 as The List's license?</h2> <p>In addition to The List mobile app, there's a web app in development too. We based the web app and the processing code on <a href="https://www.gnu.org/software/gnufm/">GNU FM</a>, a project I started back in 2009 that powers music communities such as <a href="https://libre.fm/">Libre.fm</a>. That code is under the AGPL as well, and it's a code base I am intimately familiar with. So much so that we're using it for another project too: a new federated search project at Creative Commons.</p> <h2>How can users (technical or otherwise) help contribute to The List?</h2> <p>The first thing you can do is if you have an Android phone, come try one of the public beta releases on the website, <a href="https://thelist.creativecommons.org">https://thelist.creativecommons.org</a>.</p> <p>If you're good at Android programming, you'll find our Android app in our source code under <code>app/</code>.</p> <p>If you're look at PHP, look under <code>webapp/</code>.</p> <p>And if you'd like to make some improvements to our website, they're up there too under <code>www/</code>.</p> <p>We have a really simple contributor agreement up there too. And we licensed that under CC0, if you'd like to use it for your own project.</p> <p>All of this and more can be found at <a href="https://github.com/creativecommons/list">https://github.com/creativecommons/list</a>.</p> <h2>What's the next big thing for The List?</h2> <p>The next big thing will be a proper public release. We're already talking to the F-Droid folks, and we'll be in all the places you normally find apps for your phone. And F-Droid will have the pure, free software experience.</p> <p><em>Enjoyed this interview? Check out our previous entry in this series, featuring <a href="http://www.fsf.org/blogs/licensing/the-licensing-and-compliance-lab-interviews-rainey-reitman-of-the-electronic-frontier-foundation">Rainey Reitman, Activism Director for the Electronic Frontier Foundation, about their new EFF Alerts mobile app</a>.</em></p>
]]>
</content:encoded>
<dc:publisher>No publisher</dc:publisher>
<dc:creator/>
<dc:rights/>
<dc:date>2015-04-20T16:00:00Z</dc:date>
<dc:type>Blog Entry</dc:type>
</item>
<item rdf:about="http://www.fsf.org/blogs/licensing/latest-tpp-leak-shows-systemic-threat-to-software-freedom">
<title>
Latest TPP leak shows systemic threat to software freedom
</title>
<link>
http://www.fsf.org/blogs/licensing/latest-tpp-leak-shows-systemic-threat-to-software-freedom
</link>
<description>
Key congressional leaders have just agreed on a deal to fast track the fast-tracking of TPP. While the threat of TPP has persisted for years, now is the time to fight back!
</description>
<content:encoded xmlns:content="http://purl.org/rss/1.0/modules/content/">
<![CDATA[
<p>On March 25th, 2015, Wikileaks released a leaked chapter of the ongoing Trans-Pacific Partnership (TPP) negotiations, the multinational trade agreement that is being developed through a series of secret negotiations and aims to create a host of new restrictions. We here at the FSF have been fighting against TPP for years, as it represents the threat of a world dominated by DRM, software patents, and perpetual copyright.</p> <p>The latest leaked chapter on investments lays out changes to a system of supra-national courts known extrajudicial investor-state dispute settlement (ISDS) tribunals. For years, these courts have enabled large companies to sue democratically-elected governments over policies that these corporations oppose. For example, <a href="https://www.techdirt.com/articles/20150216/17390930032/john-oliver-highlights-ridiculousness-corporate-sovereignty-provisions.shtml">Big Tobacco</a> has used the system to block or obstruct health laws intended to reduce smoking in countries around the world.</p> <p>While all of this is bad news in general, one provision in the leaked document presents a particular threat to software freedom. Holders of copyright, patent, and other proprietary interests are now included in the definition of "investor." Given the destructive nature of these provisions,